requirements.txt
bs4~=0.0.1
beautifulsoup4~=4.11.1
pandas~=1.5.1
requests~=2.28.1
yfinance~=0.1.85
dash~=2.7.0
plotly~=5.11.0
stockstats~=0.4.1
Flask~=2.2.2
python-dateutil~=2.8.2